consolidate logging and utility functions into respective modules. tidy up main, and config. add logging to api_interface
